Refactor: Update eks-node-group.yaml for sistema-gaturro-eks-ng
This commit updates the eks-node-group.yaml file for the sistema-gaturro-eks-ng node group in the eks.aws.upbound.io/v1beta1 API version. The changes include modifying the desiredSize, maxSize, and minSize values to 1, updating the subnetIdRefs to sistema-gaturro-private-subnet-b and sistema-gaturro-private-subnet-c, and setting the instanceTypes to t3.micro. These updates ensure the proper configuration of the node group in the sistema-gaturro-eks-cluster and the associated eks.aws.upbound.io provider.