Refactor: Bind AWS's IAM with Kubernetes's RBAC for EKS
This commit adds the necessary configuration to bind AWS's IAM with Kubernetes's RBAC in the eks/aws.upbound.io/v1beta1 API version. It includes modifications to the 11-openid-connect-provider.yaml file and eks.tf file. This binding allows Kubernetes to authenticate AWS users, improving the security and access control of the EKS cluster.